Refactor few code
