+++ config/alias.zsh
@@ -1,4 +1,4 @@
 alias run='./cub3D map/test.cub'
 alias norm='norminette includes srcs lib/debug > norm_result.txt'
-alias ctags_cub3d='ctags srcs/**/*.c includes/*.h lib/debug/includes/*.h lib/debug/srcs/*.c lib/libft/*.c'
+alias ctags_cub3d='ctags srcs/**/*.c lib/debug/srcs/*.c lib/libft/*.c includes/*.h lib/debug/includes/*.h'
 alias leak_check='valgrind --leak-check=full ./cub3D map/test.cub'
