+++ srcs/init_utils/parse_cubfile_utils/is_enable_map.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   is_enable_map.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kamitsui <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 13:51:35 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/10/17 19:29:51 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/10/20 02:22:46 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
