+++ srcs/render.c
@@ -6,27 +6,12 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/26 01:57:48 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/10/20 01:23:54 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/10/20 02:35:00 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void	debug_moved_player(t_game *game)
-{
-	int			fd;
-	t_vector	view_point;
-	t_vector	ray_dir;
-
-	if (IS_DEBUG == false || game->frame.flag == 0x00)
-		return ;
-	fd = STDOUT_FILENO;
-	view_point = game->player.view_point;
-	ray_dir = game->player.ray_dir;
-	dprintf(fd,
-		"moved player=<%f, %f> ray_dir=<%f, %f>\n",
-		view_point.x, view_point.y, ray_dir.x, ray_dir.y);
-}
 static void	update_player(t_game *game)
 {
 	static t_moving_player	func[6] = {
@@ -41,15 +26,13 @@ static void	update_player(t_game *game)
 	{
 		bit = 0x01 << i;
 		if (is_hit_flag(game->frame.flag, bit) == true)
-		{
-			//debug_is_hit_flag(game, bit);
 			func[i](&game->map, &game->player);
-		}
 		i++;
 	}
 	debug_moved_player(game);
 	game->frame.flag = 0;
 }
+// debug_is_hit_flag(game, bit);// call this if is_hit_flag() == true
 
 /**
  * @brief Draw the 3D perspective
